修复IdChunk获取id时int溢出

+++ idcenter-web/src/main/java/org/antframework/idcenter/web/id/core/IdChunk.java
@@ -41,7 +41,7 @@ public final class IdChunk {
         int gotAmount = Math.min(amount, this.amount);
         IdSegment idSegment = new IdSegment(period, factor, startId, gotAmount);
 
-        startId += factor * gotAmount;
+        startId += factor * (long) gotAmount;
         this.amount -= gotAmount;
 
         return idSegment;
